requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
bs4
pypdf
chromadb
faiss-cpu
groq
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
